Refactor and document
The main function offered is

    highlight(
      data, -- [{ buf_nr, line_nr, col_start, col_end, value}]
      generate_mapping_fn, -- fn(imin, imax, omin, omax) -> fn(val)
      palette,             -- [{ val: [0.0..1.0] , color: {R,G,B}}]
      highlight_namespace_name, -- string
    )

The idea is that now anybody who has a way to find the stuff (e.g.
timestamps) that they're looking for and a way to quantify a value
for them (e.g. for a timestamp, their value in seconds or something)
they can use this to get the highlighting.

Some further improvements, see doc/heat.txt